requirements.txt
numpy==2.2.1
pygame-ce==2.5.2
gymnasium==1.2.0
matplotlib==3.10.0
